pyproject.toml
[project]
name = "vectorless-rag"
version = "0.1.0"
description = "Vectorless RAG Application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "python-dotenv",
    "openai",
    "pageindex"
]
